// codigo/src/app/services/api/model/rolDto.ts
/**
 * API.Srv-inventory
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RolDto { 
    idRol?: number;
    name?: string | null;
    description?: string | null;
    estado?: string | null;
}